Signed_Operator.c
Changed variable type from signed int to signed short int.

+++ Lab_04/Signed_Operator.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main ()
 {
-   signed int a ;
+   signed short int a ;
    printf("Enter the of A : ", a );
    scanf("%u", &a ) ;
    printf("The singned value of A is : %u ", a);
